[bug-63344] add TempFilePOIFSFileSystem
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1857462 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/poifs/filesystem/TempFilePOIFSFileSystem.java
@@ -26,7 +26,7 @@ public class TempFilePOIFSFileSystem extends POIFSFileSystem {
     }
 
     public void close() throws IOException {
-        tempFile.delete();
+        if (tempFile != null) tempFile.delete();
         super.close();
     }
